pyproject.toml
[project]
name = "eve-esi-tool"
version = "0.1.0"
description = "EVE Online ESI API tool with MCP server for AI agent integration"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "cryptography>=41.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "mcp[cli]>=1.0.0",
]

[project.scripts]
eve-esi = "cli:main"
eve-mcp = "mcp_server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

